Redesign interrupt handler stuff
Instead of having the same code duplicated 256 times for interrupts,
have an interrupt table that stores the pointers to each interrupt
handler (in Assembly). Then, in C when setting up the IDT, use that
instead of calculating the address of the handler.

This way macros can be used to only include code if it's necessary,
like popping the error code from the stack when there is an error code.

Also removed an unnecessary `pic_sendEOI`.