Usermode jump code
Doesn't work yet for multiple reasons:

- TSS is not filled in properly before entering user mode
- The "process" doesn't have a separate stack so even if the TSS RSP0
was set before entering user mode, interrupts would overwrite the
"process"'s stack
- The "process" does not have its own page table with the user bit set,
so it `#PF`s whenever it tries to do basically anything. In the current
state this results in a triple fault because of the TSS as mentioned
above (namely CR3 is set to 0 which results in another `#PF`…).

I solemnly swear that I will not try to jump to usermode again before
I have a process/task system set up. That is my next task (no pun
intended).